SYNOPSIS pod init_arg correction
Note that our init_arg for the constructed class attribute (thinger_class in
this case) is set to undef, not the original class attribute.  The original
class attribute's init_arg is set to the name of the constructed class
attribute (e.g. original_thinger_class' init_arg is 'thinger_class').

This allows things to work, as expected, without worrying too much.