Prevent our exports from being assimilated as role methods
Using Moose::Exporter in a role had the unanticipated consequence that
the functions exported via with_meta are also picked up by the role
metaclass as methods.

Oops.

So this is a little awkward, but it resolves this unpleasantness.

Here we simply create a "dummy" package using Moose exporter, then pull
its imports in via an "also" in the main package.  This resolves things
nicely without having to do any meta-tinkering.

Sadly.